+++ lib/routes/transaction_page/sections/tags_section.dart
@@ -1,65 +1,95 @@
+import "package:flow/entity/transaction/extensions/default/geo.dart";
 import "package:flow/entity/transaction_tag.dart";
 import "package:flow/l10n/extensions.dart";
 import "package:flow/prefs/local_preferences.dart";
+import "package:flow/providers/transaction_tags_provider.dart";
 import "package:flow/routes/transaction_page/section.dart";
-import "package:flow/widgets/general/directional_chevron.dart";
 import "package:flow/widgets/general/frame.dart";
 import "package:flow/widgets/general/info_text.dart";
+import "package:flow/widgets/transaction_tag_add_chip.dart";
 import "package:flow/widgets/transaction_tag_chip.dart";
 import "package:flutter/material.dart";
 import "package:flutter/services.dart";
-import "package:material_symbols_icons/symbols.dart";
+import "package:latlong2/latlong.dart";
 
 class TagsSection extends StatelessWidget {
   final List<TransactionTag>? selectedTags;
   final VoidCallback selectTags;
+  final ValueChanged<List<TransactionTag>> onTagsChanged;
 
-  const TagsSection({super.key, this.selectedTags, required this.selectTags});
+  /// Used for suggesting nearby tags based on the transaction's location.
+  final Geo? location;
+
+  const TagsSection({
+    super.key,
+    this.selectedTags,
+    required this.selectTags,
+    required this.onTagsChanged,
+    this.location,
+  });
 
   @override
   Widget build(BuildContext context) {
+    final List<TransactionTag>? suggestedGeoTags = switch (location
+        ?.toLatLngPosition()) {
+      LatLng latLng => TransactionTagsProvider.of(
+        context,
+      ).getCloseGeoTags(latLng, exclusionList: selectedTags),
+      _ => null,
+    };
+
+    final bool hasSuggestedGeoTags = suggestedGeoTags?.isNotEmpty == true;
+
     return Section(
       title: "transaction.tags".t(context),
       child: GestureDetector(
         behavior: HitTestBehavior.opaque,
-        child: selectedTags?.isNotEmpty == true
-            ? Frame(
-                child: Padding(
-                  padding: const EdgeInsets.only(top: 8.0),
-                  child: Align(
-                    alignment: AlignmentDirectional.topStart,
-                    child: Column(
-                      crossAxisAlignment: .start,
-                      spacing: 8.0,
-                      children: [
-                        IgnorePointer(
-                          child: Wrap(
-                            spacing: 12.0,
-                            runSpacing: 8.0,
-                            children: selectedTags!
-                                .map(
-                                  (tag) => TransactionTagChip(
-                                    tag: tag,
-                                    selected: true,
-                                  ),
-                                )
-                                .toList(),
+        child: Frame(
+          child: Padding(
+            padding: const EdgeInsets.only(top: 8.0),
+            child: Align(
+              alignment: AlignmentDirectional.topStart,
+              child: Column(
+                crossAxisAlignment: .start,
+                spacing: 8.0,
+                children: [
+                  Wrap(
+                    spacing: 12.0,
+                    runSpacing: 8.0,
+                    children: [
+                      TransactionTagAddChip(
+                        onPressed: selectTags,
+                        title: "transaction.tags.add".t(context),
+                      ),
+                      if (hasSuggestedGeoTags)
+                        ...suggestedGeoTags!.map(
+                          (tag) => TransactionTagChip(
+                            tag: tag,
+                            selected: false,
+                            isSuggestion: true,
+                            onPressed: () {
+                              _addTag(context, tag);
+                            },
                           ),
                         ),
-                        InfoText(
-                          child: Text("transaction.tags.editGuide".t(context)),
+                      ...selectedTags!.map(
+                        (tag) => IgnorePointer(
+                          child: TransactionTagChip(tag: tag, selected: true),
                         ),
-                      ],
-                    ),
+                      ),
+                    ],
                   ),
-                ),
-              )
-            : ListTile(
-                leading: Icon(Symbols.style_rounded),
-                title: Text("transaction.edit.selectTags".t(context)),
-                trailing: LeChevron(),
-                onTap: selectTags,
+                  if (hasSuggestedGeoTags)
+                    InfoText(
+                      child: Text(
+                        "transaction.tags.suggestionGuide".t(context),
+                      ),
+                    ),
+                ],
               ),
+            ),
+          ),
+        ),
         onTap: () {
           if (LocalPreferences().enableHapticFeedback.get()) {
             HapticFeedback.lightImpact();
@@ -69,5 +99,15 @@ class TagsSection extends StatelessWidget {
         },
       ),
     );
+  }
+
+  void _addTag(BuildContext context, TransactionTag tag) {
+    if (selectedTags!.contains(tag)) return;
+
+    if (LocalPreferences().enableHapticFeedback.get()) {
+      HapticFeedback.lightImpact();
+    }
+
+    onTagsChanged([...selectedTags!, tag]);
   }
 }