+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Added a Siri integration (intent).
 * Persian (Iran) support thanks to @arefbhrn[https://github.com/arefbhrn]
+* Location tags are suggested if you're within 50m of the tag. Requires location
+  to be enabled. Closes [#648](https://github.com/flow-mn/flow/issues/648)
 
 ### Changes